+++ src/components/Tabs.vue
@@ -38,6 +38,7 @@
                 required: false,
                 default: () => ({
                     useUrlFragment: true,
+                    defaultTabHash: null,
                 }),
             },
         },
@@ -45,6 +46,8 @@
         data: () => ({
             tabs: [],
             activeTabHash: '',
+            activeTabIndex: 0,
+            lastActiveTabHash: '',
         }),
 
         computed: {
@@ -72,6 +75,11 @@
                 return;
             }
 
+            if(this.options.defaultTabHash !== null && this.findTab("#" + this.options.defaultTabHash)) {
+                this.selectTab("#" + this.options.defaultTabHash);
+                return;
+            }
+
             if (this.tabs.length) {
                 this.selectTab(this.tabs[0].hash);
             }
@@ -95,6 +103,12 @@
                 }
 
                 if (selectedTab.isDisabled) {
+                    event.preventDefault();
+                    return;
+                }
+
+                if (this.lastActiveTabHash === selectedTab.hash) {
+                    this.$emit('clicked', { tab: selectedTab });
                     return;
                 }
 
@@ -105,6 +119,9 @@
                 this.$emit('changed', { tab: selectedTab });
 
                 this.activeTabHash = selectedTab.hash;
+                this.activeTabIndex = this.getTabIndex(selectedTabHash);
+
+                this.lastActiveTabHash = this.activeTabHash = selectedTab.hash;
 
                 expiringStorage.set(this.storageKey, selectedTab.hash, this.cacheLifetime);
             },
@@ -132,6 +149,30 @@
                         return true;
                     });
                 }
+            },
+            
+            getTabIndex(hash){
+            	const tab = this.findTab(hash);
+            	
+            	return this.tabs.indexOf(tab);
+            },
+            
+			getTabHash(index){
+            	const tab = this.tabs.find(tab => this.tabs.indexOf(tab) === index);
+            	
+            	if (!tab) {
+					return;
+                }
+                
+                return tab.hash;
+			},
+            
+            getActiveTab(){
+            	return this.findTab(this.activeTabHash);
+            },
+            
+			getActiveTabIndex() {
+            	return this.getTabIndex(this.activeTabHash);
             },
         },
     };