+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 All notable changes to `vue-tabs-component` will be documented in this file
 
-## 1.4.1 - 2018-03-06
+## 1.5.0 - 2018-XX-XX
+- Added `defaultTabHash` option
+- Fixed bug #32 `changed` event fires twice on each change
+- Added `clicked` event, fires when active tab is clicked
 - Fixed a bug which caused sections within tabs to have a '#' included in their id attribute.
 
 ## 1.4.0 - 2017-11-06